requirements.txt
argparse
hydra-core
tensorflow
numpy
matplotlib
kaggle
zipfile